+++ src/Contracts/Schematized.php
@@ -20,5 +20,5 @@ interface Schematized
 {
     public function getSchema(): Schema;
 
-    public function getSchemaSample(array $mergeWith = null): array;
+    public function getSchemaSample(?array $mergeWith = null): array;
 }