+++ src/Payment/Contracts/PaymentGateway.php
@@ -26,7 +26,7 @@ interface PaymentGateway
 
     public function createPaymentRequest(
         Payment $payment,
-        Address $shippingAddress = null,
+        ?Address $shippingAddress = null,
         array $options = []
     ): PaymentRequest;