+++ src/Payment/Events/BasePaymentEvent.php
@@ -14,12 +14,14 @@ declare(strict_types=1);
 
 namespace Vanilo\Payment\Events;
 
+use Illuminate\Queue\SerializesModels;
 use Vanilo\Payment\Contracts\Payment;
 use Vanilo\Payment\Contracts\PaymentEvent;
 
 class BasePaymentEvent implements PaymentEvent
 {
     use HasPayment;
+    use SerializesModels;
 
     public function __construct(Payment $payment)
     {