+++ src/Payment/Changelog.md
@@ -6,6 +6,7 @@
 ##### 2025-XX-XX
 
 - Dropped PHP 8.2 Support
+- Changed the minimum Laravel 10 version to v10.48
 - Added Laravel 12 Support
 - PHP 8.4 deprecation notice fixes:
    - The `$shippingAddress` parameter in the `PaymentGateway::createPaymentRequest()` interface method is now explicitly nullable